# backend/src/config/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Delora Market API",
        "version": "1.0.0",
        "description": "API documentation for Delora Hypermarket backend"
    },
    "servers": [
        { "url": "http://localhost:5000", "description": "Local dev" }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    },
    "security": [{ "BearerAuth": [] }],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": ["Auth"],
                "summary": "Register new user",
                "security": [],
                "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "email": { "type": "string" }, "password": { "type": "string" }, "phone": { "type": "string" }, "gender": { "type": "string", "enum": ["male","female","other"] } }, "required": ["name","email","password","phone"] } } } },
                "responses": { "201": { "description": "User created, OTP sent" }, "400": { "description": "Validation error / email exists" } }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": ["Auth"], "security": [],
                "summary": "Login (returns JWT)",
                "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string" }, "password": { "type": "string" } }, "required": ["email","password"] } } } },
                "responses": { "200": { "description": "Login success + JWT" }, "403": { "description": "Email not verified" } }
            }
        },
        "/api/auth/verify-email": {
            "post": {
                "tags": ["Auth"], "security": [],
                "summary": "Verify email with OTP",
                "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string" }, "otp": { "type": "string" } }, "required": ["email","otp"] } } } },
                "responses": { "200": { "description": "Email verified" } }
            }
        },
        "/api/auth/resend-otp": {
            "post": {
                "tags": ["Auth"], "security": [],
                "summary": "Resend verification OTP",
                "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string" } }, "required": ["email"] } } } },
                "responses": { "200": { "description": "OTP resent" } }
            }
        },
        "/api/auth/forgot-password": {
            "post": {
                "tags": ["Auth"], "security": [],
                "summary": "Send password reset OTP",
                "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string" } }, "required": ["email"] } } } },
                "responses": { "200": { "description": "Reset OTP sent" } }
            }
        },
        "/api/auth/reset-password": {
            "post": {
                "tags": ["Auth"], "security": [],
                "summary": "Reset password with OTP",
                "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string" }, "otp": { "type": "string" }, "new_password": { "type": "string" } }, "required": ["email","otp","new_password"] } } } },
                "responses": { "200": { "description": "Password reset" } }
            }
        },
        "/api/auth/profile": {
            "get": {
                "tags": ["Auth"], "summary": "Get current user profile",
                "responses": { "200": { "description": "User profile" } }
            },
            "put": {
                "tags": ["Auth"], "summary": "Update profile (name, phone, gender)",
                "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "phone": { "type": "string" }, "gender": { "type": "string" } } } } } },
                "responses": { "200": { "description": "Profile updated" } }
            }
        },
...
    }
}